requirements.txt
anthropic>=0.40.0
sqlalchemy>=2.0
psycopg2-binary>=2.9
pandas>=2.2
streamlit>=1.40
python-dotenv>=1.0
pyyaml>=6.0
structlog>=24.0
